+++ messages/ko/forms_CreateMessageForm.php
@@ -1,8 +1,7 @@
 <?php
-
-return [
-    'Message' => '',
-    'Recipient' => '',
-    'Subject' => '',
-    'Tags' => '',
-];
+return array (
+  'Message' => '',
+  'Recipient' => '',
+  'Subject' => '',
+  'Tags' => 'ㅡㅡ',
+);