+++ messages/lt/forms_InviteRecipientForm.php
@@ -1,8 +1,7 @@
 <?php
-
-return [
-    'User {name} is already participating!' => '',
-    'You are not allowed to send user {name} is already!' => '',
-    'You cannot send a email to yourself!' => '',
-    'Recipient' => 'Gavėjas',
-];
+return array (
+  'Recipient' => 'Gavėjas',
+  'User {name} is already participating!' => 'Vartotojas {name} jau dalyvauja!',
+  'You are not allowed to send user {name} is already!' => 'Jums neleidžiama siųsti vartotojo {name} jau yra!',
+  'You cannot send a email to yourself!' => 'Negalite sau siųsti el. laiško!',
+);