+++ messages/sk/views_emails_NewMessage.php
@@ -1,7 +1,6 @@
 <?php
-
-return [
-    '<strong>New</strong> conversation' => '',
-    '<strong>New</strong> message' => '',
-    'Reply now' => '',
-];
+return array (
+  '<strong>New</strong> conversation' => '<strong>Nová</strong> konverzácia',
+  '<strong>New</strong> message' => '<strong>Nová</strong> správa',
+  'Reply now' => 'Odpovedať teraz',
+);