+++ messages/sk/views_emails_NewMessageEntry.php
@@ -1,6 +1,5 @@
 <?php
-
-return [
-    '{senderName} created a new conversation {conversationTitle}' => '',
-    '{senderName} sent you a new message in {conversationTitle}' => '',
-];
+return array (
+  '{senderName} created a new conversation {conversationTitle}' => '{senderName} vytvoril novú konverzáciu {conversationTitle}',
+  '{senderName} sent you a new message in {conversationTitle}' => '{senderName} vám poslal novú správu v {conversationTitle}',
+);