+++ messages/sk/views_mail_show.php
@@ -1,18 +1,17 @@
 <?php
-
-return [
-    '<strong>Confirm</strong> deleting conversation' => '',
-    '<strong>Confirm</strong> leaving conversation' => '',
-    '<strong>Confirm</strong> message deletion' => '',
-    'Add user' => '',
-    'Cancel' => '',
-    'Delete' => '',
-    'Delete conversation' => '',
-    'Do you really want to delete this conversation?' => '',
-    'Do you really want to delete this message?' => '',
-    'Do you really want to leave this conversation?' => '',
-    'Leave' => '',
-    'Leave conversation' => '',
-    'There are no messages yet.' => '',
-    'You are not allowed to participate in this conversation. You have been blocked by: {userNames}.' => '',
-];
+return array (
+  '<strong>Confirm</strong> deleting conversation' => '<strong>Potvrďte</strong> odstránenie konverzácie',
+  '<strong>Confirm</strong> leaving conversation' => '<strong>Potvrďte</strong> opustenie konverzácie',
+  '<strong>Confirm</strong> message deletion' => '<strong>Potvrďte</strong> odstránenie správy',
+  'Add user' => 'Pridať používateľa',
+  'Cancel' => 'Zrušiť',
+  'Delete' => 'Odstrániť',
+  'Delete conversation' => 'Odstrániť konverzáciu',
+  'Do you really want to delete this conversation?' => 'Naozaj chcete odstrániť túto konverzáciu?',
+  'Do you really want to delete this message?' => 'Naozaj chcete odstrániť túto správu?',
+  'Do you really want to leave this conversation?' => 'Naozaj chcete opustiť túto konverzáciu?',
+  'Leave' => 'Odísť',
+  'Leave conversation' => 'Opustiť konverzáciu',
+  'There are no messages yet.' => 'Zatiaľ nie sú žiadne správy.',
+  'You are not allowed to participate in this conversation. You have been blocked by: {userNames}.' => 'Nemáte povolené zúčastniť sa tejto konverzácie. Boli ste zablokovaní používateľom: {userNames}.',
+);