+++ messages/sr/models_Message.php
@@ -1,6 +1,5 @@
 <?php
-
-return [
-    'New conversation from {senderName}' => '',
-    'New message from {senderName}' => '',
-];
+return array (
+  'New conversation from {senderName}' => '',
+  'New message from {senderName}' => 'Nova poruka od {senderName}',
+);