+++ models/MessageNotification.php
@@ -4,11 +4,13 @@ namespace humhub\modules\mail\models;
 
 use humhub\modules\content\widgets\richtext\converter\RichTextToEmailHtmlConverter;
 use humhub\modules\content\widgets\richtext\converter\RichTextToHtmlConverter;
+use humhub\modules\mail\helpers\Url;
 use humhub\modules\mail\live\NewUserMessage;
 use humhub\modules\mail\notifications\ConversationNotificationCategory;
 use humhub\modules\mail\notifications\MailNotificationCategory;
 use humhub\modules\notification\components\NotificationCategory;
 use humhub\modules\notification\targets\MailTarget;
+use humhub\modules\notification\targets\MobileTarget;
 use humhub\modules\user\models\User;
 use Yii;
 use yii\base\BaseObject;
@@ -27,6 +29,11 @@ class MessageNotification extends BaseObject
     public $entry;
 
     /**
+     * @var User
+     */
+    public $entrySender;
+
+    /**
      * @var bool $isNewConversation Flag for notification type: Conversation vs Message
      */
     public $isNewConversation = false;
@@ -34,7 +41,7 @@ class MessageNotification extends BaseObject
     public function __construct(Message $message, MessageEntry $entry = null)
     {
         $this->message = $message;
-        $this->entry = $entry ?: $this->message->lastEntry;
+        $this->entry = $entry ?? $this->message->lastEntry;
         parent::__construct([]);
     }
 
@@ -55,6 +62,8 @@ class MessageNotification extends BaseObject
 
             $this->sendMail($user);
 
+            $this->sendPush($user);
+
             // Restore the flag
             $this->isNewConversation = $isNewConversation;
         } catch (\Exception $e) {
@@ -67,8 +76,35 @@ class MessageNotification extends BaseObject
         Yii::$app->live->send(new NewUserMessage([
             'contentContainerId' => $user->contentcontainer_id,
             'message_id' => $this->message->id,
-            'user_guid' => $user->guid
+            'user_guid' => $user->guid,
         ]));
+    }
+
+    private function canReceiveByTarget(User $user, string $targetClass): bool
+    {
+        if ($user->status != User::STATUS_ENABLED) {
+            return false;
+        }
+
+        if ($user->is($this->getEntrySender())) {
+            return false;
+        }
+
+        if (!($target = Yii::$app->notification->getTarget($targetClass))) {
+            return false;
+        }
+
+        if ($target->isCategoryEnabled($this->getNotificationCategory(), $user)) {
+            return true;
+        }
+
+        // Try to send notification as "New message" when notification "New conversation" is disabled for the user
+        if ($this->isNewConversation && $target->isCategoryEnabled(new MailNotificationCategory(), $user)) {
+            $this->isNewConversation = false;
+            return true;
+        }
+
+        return false;
     }
 
     private function canReceiveMail(User $user): bool
@@ -77,25 +113,12 @@ class MessageNotification extends BaseObject
             return false;
         }
 
-        if ($user->is($this->getEntrySender())) {
-            return false;
-        }
+        return $this->canReceiveByTarget($user, MailTarget::class);
+    }
 
-        if (!($mailTarget = Yii::$app->notification->getTarget(MailTarget::class))) {
-            return false;
-        }
-
-        if ($mailTarget->isCategoryEnabled($this->getNotificationCategory(), $user)) {
-            return true;
-        }
-
-        // Try to send notification as "New message" when notification "New conversation" is disabled for the user
-        if ($this->isNewConversation && $mailTarget->isCategoryEnabled(new MailNotificationCategory(), $user)) {
-            $this->isNewConversation = false;
-            return true;
-        }
-
-        return false;
+    private function canReceivePush(User $user): bool
+    {
+        return $this->canReceiveByTarget($user, MobileTarget::class);
     }
 
     private function getNotificationCategory(): NotificationCategory
@@ -115,7 +138,7 @@ class MessageNotification extends BaseObject
 
         $mail = Yii::$app->mailer->compose([
             'html' => '@mail/views/emails/NewMessage',
-            'text' => '@mail/views/emails/plaintext/NewMessage'
+            'text' => '@mail/views/emails/plaintext/NewMessage',
         ], [
             'user' => $user,
             'headline' => $this->getHeadline(),
@@ -130,14 +153,42 @@ class MessageNotification extends BaseObject
 
         $mail->setFrom([Yii::$app->settings->get('mailer.systemEmailAddress') => Yii::$app->settings->get('mailer.systemEmailName')]);
         $mail->setTo($user->email);
-        $mail->setSubject($this->getSubject());
+        $mail->setSubject($this->getSubject($user));
         $mail->send();
 
         Yii::$app->i18n->autosetLocale();
     }
 
+    private function sendPush(User $user)
+    {
+        $fcmModule = Yii::$app->getModule('fcm-push');
+        if (!$fcmModule || !$fcmModule->isActivated) {
+            return;
+        }
+        if (!$this->canReceivePush($user)) {
+            return;
+        }
+
+        $firebaseService = new \humhub\modules\fcmPush\services\MessagingService($fcmModule->getConfigureForm());
+
+        $firebaseService->processMessage(
+            $user,
+            Yii::$app->name,
+            $this->getSubHeadline(),
+            Url::toMessenger($this->message, true),
+            null,
+            null,
+        );
+    }
+
     protected function getContent(User $user)
     {
+        if ($this->entry->type === AbstractMessageEntry::TYPE_USER_JOINED) {
+            return $this->entry->user->is($user)
+                ? Yii::t('MailModule.base', 'You joined the conversation.')
+                : Yii::t('MailModule.base', '{username} joined the conversation.', ['username' => $this->entry->user->displayName]);
+        }
+
         return RichTextToEmailHtmlConverter::process($this->entry->content, [
             RichTextToEmailHtmlConverter::OPTION_RECEIVER_USER => $user,
             RichTextToHtmlConverter::OPTION_LINK_AS_TEXT => true,
@@ -172,17 +223,29 @@ class MessageNotification extends BaseObject
         return $this->message->originator;
     }
 
+    public function setEntrySender(User $user): self
+    {
+        $this->entrySender = $user;
+        return $this;
+    }
+
     /**
      * @return User
      */
     protected function getEntrySender()
     {
-        return $this->entry->user;
+        return $this->entrySender ?? $this->entry->user;
     }
 
-    protected function getSubject(): string
+    protected function getSubject(User $user): string
     {
         $params = ['{senderName}' => $this->getEntrySender()->displayName];
+
+        if ($this->entry->type === AbstractMessageEntry::TYPE_USER_JOINED) {
+            return $this->entry->user->is($user)
+                ? Yii::t('MailModule.base', 'You joined the conversation.')
+                : Yii::t('MailModule.base', '{username} joined the conversation.', ['username' => $this->entry->user->displayName]);
+        }
 
         return $this->isNewConversation
             ? Yii::t('MailModule.models_Message', 'New conversation from {senderName}', $params)