+++ tests/codeception/unit/NotificationTest.php
@@ -37,7 +37,7 @@ class NotificationTest extends HumHubDbTestCase
             'title' => $title,
             'message' => $message,
             'recipient' => $users,
-            'tags' => $tags
+            'tags' => $tags,
         ]);
 
         $this->assertTrue($message->save());
@@ -73,7 +73,7 @@ class NotificationTest extends HumHubDbTestCase
         $this->assertEqualsLastEmailSubject('New conversation from Peter Tester');
         $this->assertEqualsLastEmailTo($user2->email);
         $test = Live::find()->all();
-        $this->assertCount(1,  Live::find()->where(['contentcontainer_id' => $user2->contentcontainer_id])->all());
+        $this->assertCount(1, Live::find()->where(['contentcontainer_id' => $user2->contentcontainer_id])->all());
     }
 
     public function testNotificationsAreSentAllRecepients()
@@ -92,21 +92,23 @@ class NotificationTest extends HumHubDbTestCase
         $test =  Live::find(['contentcontainer_id' => $user2->contentcontainer_id])->all();
         $test2 = unserialize($test[0]->serialized_data);
         $test3 = unserialize($test[1]->serialized_data);
-        $this->assertCount(1,  Live::find()->where(['contentcontainer_id' => $user2->contentcontainer_id])->all());
-        $this->assertCount(1,  Live::find()->where(['contentcontainer_id' => $user3->contentcontainer_id])->all());
+        $this->assertCount(1, Live::find()->where(['contentcontainer_id' => $user2->contentcontainer_id])->all());
+        $this->assertCount(1, Live::find()->where(['contentcontainer_id' => $user3->contentcontainer_id])->all());
     }
 
     public function testNotificationsAreSentAfterAddingParticipant()
     {
         $user2 = User::findOne(['id' => 3]);
         $message = $this->createMessage('First', 'First', [$user2->guid]);
-
+        $this->assertSentEmail(1);
+        $this->assertEqualsLastEmailSubject('New conversation from Peter Tester');
+        $this->assertEqualsLastEmailTo($user2->email);
 
         $user3 = User::findOne(['id' => 4]);
         $inviteForm = new InviteParticipantForm(['message' => $message->messageInstance, 'recipients' => [$user3->guid]]);
         $inviteForm->save();
 
-        $this->assertSentEmail(2);
+        $this->assertSentEmail(3);
         $this->assertEqualsLastEmailSubject('New message from Peter Tester');
         $this->assertEqualsLastEmailTo($user3->email);
     }