+++ views/emails/plaintext/NewMessage.php
@@ -4,8 +4,6 @@ use humhub\modules\mail\helpers\Url;
 use humhub\modules\mail\models\Message;
 use humhub\modules\mail\models\MessageEntry;
 use humhub\modules\user\models\User;
-use yii\helpers\Html;
-use humhub\widgets\MarkdownView;
 
 /* @var $user User */
 /* @var $sender User */