+++ widgets/ConversationTags.php
@@ -1,8 +1,6 @@
 <?php
 
-
 namespace humhub\modules\mail\widgets;
-
 
 use humhub\components\Widget;
 use humhub\libs\Html;
@@ -11,19 +9,20 @@ use Yii;
 
 class ConversationTags extends Widget
 {
-    const ID = 'conversation-tags-root';
+    public const ID = 'conversation-tags-root';
 
     public $message;
 
     public function run()
-    {   $tags = MessageTag::findByMessage(Yii::$app->user->id, $this->message)->all();
+    {
+        $tags = MessageTag::findByMessage(Yii::$app->user->id, $this->message)->all();
 
         $result = Html::beginTag('div', ['id' => static::ID, 'class' => 'panel-body', 'style' => ['display' => count($tags) ? 'block' : 'none']]);
 
-        $result .= '<span class="my-tags-label">'.Yii::t('MailModule.base', 'My Tags').'</span>';
+        $result .= '<span class="my-tags-label">' . Yii::t('MailModule.base', 'My Tags') . '</span>';
 
         foreach ($tags as $tag) {
-            $result .= ConversationTagBadge::get($tag).'&nbsp;';
+            $result .= ConversationTagBadge::get($tag) . '&nbsp;';
         }
 
         $result .= ConversationTagBadge::getEditConversationTagBadge($this->message, (empty($tags) ? 'plus' : 'pencil'));